requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
python-telegram-bot==21.4
lxml==5.2.2
